aimentor-ai/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1